requirements.txt
alembic==1.13.2
anyio==3.7.1
aiohttp==3.11.2
asyncpg==0.29.0
bcrypt==4.1.3
cryptography==42.0.4
environs==11.0.0
fastapi==0.115.5
fastapi-cli==0.0.4
Jinja2==3.1.4
Mako==1.3.5
pydantic==2.7.4
pydantic-settings==2.3.4
pydantic_core==2.18.4
PyJWT==2.8.0
pytest~=8.3.3
pytest-asyncio==0.24.0
pytest-dotenv==0.5.2
python-dotenv==1.0.1
SQLAlchemy==2.0.31
uvicorn~=0.30.1
pytest~=8.3.3
httpx~=0.27.2